consistent job and publish connector package names
