+++ mantis-connector-job/build.gradle
