+++ mantis-connector-publish/build.gradle
